requirements.txt
flask
pillow
numpy
opencv-python
rembg
Flask==2.3.3
Pillow==10.0.1
rembg==2.0.34  # optional if background removal is needed
numpy==1.26.4
